+++ forge/src/main/java/com/gitlab/srcmc/rctmod/forge/CobblemonHandler.java
@@ -103,6 +103,17 @@ public class CobblemonHandler {
         if(owner != null) {
             var playerTr = RCTMod.get().getTrainerManager().getData(owner);
             var maxExp = event.getPokemon().getExperienceToLevel(playerTr.getLevelCap());
+
+            if(maxExp < event.getExperience()) {
+                owner.server.getCommands().performPrefixedCommand(
+                    owner.server.createCommandSourceStack().withSuppressedOutput(),
+                    String.format("title %s actionbar \"%s is %s the level cap (%d)\"",
+                        owner.getName().getString(),
+                        event.getPokemon().getDisplayName().getString(),
+                        event.getPokemon().getLevel() == playerTr.getLevelCap() ? "at" : "over",
+                        playerTr.getLevelCap()));
+            }
+
             event.setExperience(Math.min(event.getExperience(), maxExp));
         }